-netbsd/64 fixes and warnings fixes

+++ modules_s/auth_diameter/user_in.c
@@ -128,7 +128,7 @@ int diameter_is_user_in(struct sip_msg* _m, char* _hf, char* _group)
 
 	grp = (str*)_group; /* via fixup */
 
-	hf_type = (int)_hf;
+	hf_type = (int)(long)_hf;
 
 	/* extract the uri according with the _hf parameter */
 	switch(hf_type)
